+++ .php-cs-fixer.dist.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+$finder = (new PhpCsFixer\Finder())
+    ->in(__DIR__ . '/src')
+    ->in(__DIR__ . '/tests');
+
+return (new PhpCsFixer\Config())
+    ->setRiskyAllowed(true)
+    ->setRules([
+        '@PSR12' => true,
+        '@PHP84Migration' => true,
+        'declare_strict_types' => true,
+        'strict_param' => true,
+        'ordered_imports' => true,
+        'no_unused_imports' => true,
+        'trailing_comma_in_multiline' => true,
+        'single_quote' => true,
+        'array_syntax' => ['syntax' => 'short'],
+    ])
+    ->setFinder($finder);